refactor: rename docs to old-docs and new docs to /docs
