--- src/pages/newDocs/index.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import { NewDocs } from 'views/NewDocs';
-
-export { default as getServerSideProps } from 'lib/accounts/protectedRouteGetServerSideProps';
-
-const NewDocsPage = () => {
-  return <NewDocs />;
-};
-
-export default NewDocsPage;
